fix: remove button rename from this PR
