+++ __tests__/components/buttons/NormalButton.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from "@testing-library/react";
-import NormalButton from "@/components/buttons/Button";
+import NormalButton from "@/components/buttons/NormalButton";
 import "@testing-library/jest-dom";
 
 /*=====================================================*/
